requirements.txt
# ── Core pipeline ─────────────────────────────────────────────────
pandas==2.0.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
pyarrow==16.0.0

# ── Machine learning ──────────────────────────────────────────────
prophet==1.3.0               # matches your installed version
lightgbm==4.3.0
scikit-learn==1.4.2
joblib==1.4.2

# ── Dashboard (Streamlit) ─────────────────────────────────────────
streamlit==1.28.2
plotly==5.18.0
streamlit-autorefresh==1.0.1

# ── Testing ───────────────────────────────────────────────────────
pytest==8.2.2
pytest-cov==5.0.0